demande fourniture

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -36,7 +36,7 @@ class CreateUsersTable extends Migration
             $table->string('email')->unique();
             $table->string('password');
             $table->string('img')->nullable();
-            $table->smallInteger('role')->default(0);
+            $table->float('role')->default(0);
 
             /* student */
